Replace vld1_dup_X with vdup_n_X
I think the vld1_* form is dangerous, because if the scalar type differs, it accesses a different amount of memory than the type indicates. And if the type matches, at least clang produces identical intermediate code anyways: https://godbolt.org/z/Gc4EK67jM

This should fix many warnings.

PiperOrigin-RevId: 704890593

+++ src/f16-avgpool/f16-avgpool-9x-minmax-neonfp16arith-c8.c
@@ -162,7 +162,7 @@ void xnn_f16_avgpool_minmax_ukernel_9x__neonfp16arith_c8(
         vout_lo = vget_high_f16(vout);
       }
       if (c & 2) {
-        vst1_lane_u32((uint16_t*) output, vreinterpret_u32_f16(vout_lo), 0); output = (xnn_float16*) output + 2;
+        vst1_lane_u32((uint32_t*) output, vreinterpret_u32_f16(vout_lo), 0); output = (xnn_float16*) output + 2;
         vout_lo = vext_f16(vout_lo, vout_lo, 2);
       }
       if (c & 1) {
